perf: 修改 仅一次 逻辑与当前一致
